refactor: Remove AudioRecorder component and associated styles; replace with updated implementation in ConversationFooter
