+++ frontend/src/components/layout/Conversation/ConversationFooter/AudioRecorder.jsx
@@ -1,75 +1,49 @@
-// AudioRecorder.jsx  
 import React, { useState, useRef, useCallback, useEffect } from 'react';  
 import PropTypes from 'prop-types';  
-import { useTranslation } from 'react-i18next';  
-import { SERVER_WEBSOCKET_URL } from '../../../../config/config';  
-import './AudioRecorder.scss';
+import { useTranslation } from "react-i18next";  
+
+import { SERVER_WEBSOCKET_URL } from '../../../../config/config';
 
 const AudioRecorder = ({ setInput, isStreaming, setError }) => {  
     const { t } = useTranslation();  
     const [isRecording, setIsRecording] = useState(false);  
     const mediaRecorderRef = useRef(null);  
-    const webSocketRef = useRef(null);  
-    const audioChunksRef = useRef([]);
+    const webSocketRef = useRef(null);
 
     const handleMessage = useCallback((event) => {  
         try {  
             const message = JSON.parse(event.data);  
             if (message.text) {  
-                setInput(prevInput => prevInput ? `${prevInput} ${message.text}` : message.text);  
+                setInput(prevInput => `${prevInput} ${message.text}`);  
             } else {  
                 setError(`Error: ${event.data}`);  
             }  
         } catch (error) {  
-            setError(`Error: ${error.message} - ${event.data}`);  
+            setError(`Error: ${error.message} ${event.data}`);  
         }  
     }, [setInput, setError]);
 
     const initializeWebSocket = useCallback(() => {  
         return new Promise((resolve, reject) => {  
-            const ws = new WebSocket(SERVER_WEBSOCKET_URL);  
-              
-            ws.addEventListener('open', () => resolve(ws));  
-            ws.addEventListener('message', handleMessage);  
-            ws.addEventListener('error', (error) => {  
-                console.error('WebSocket Error:', error);  
+            try {  
+                const ws = new WebSocket(SERVER_WEBSOCKET_URL);  
+                ws.onopen = () => resolve(ws);  
+                ws.onmessage = handleMessage;  
+                ws.onerror = reject;  
+            } catch (error) {  
                 reject(error);  
-            });  
-            ws.addEventListener('close', () => {  
-                console.log('WebSocket connection closed');  
-            });  
+            }  
         });  
     }, [handleMessage]);
 
-    const convertBlobToBase64 = useCallback((blob) => {  
-        return new Promise((resolve, reject) => {  
-            const reader = new FileReader();  
-            reader.onloadend = () => resolve(reader.result.split(',')[1]);  
-            reader.onerror = reject;  
-            reader.readAsDataURL(blob);  
-        });  
-    }, []);
 
-    const handleRecordingStop = useCallback(async (stream) => {  
-        setIsRecording(false);  
-        const audioBlob = new Blob(audioChunksRef.current, { type: 'audio/webm' });  
-        audioChunksRef.current = [];
+    const convertBlobToBase64 = useCallback((blob) => new Promise((resolve, reject) => {  
+        const reader = new FileReader();  
+        reader.onloadend = () => resolve(reader.result.split(',')[1]);  
+        reader.onerror = reject;  
+        reader.readAsDataURL(blob);  
+    }), []);
 
-        try {  
-            const base64Data = await convertBlobToBase64(audioBlob);  
-            if (webSocketRef.current?.readyState === WebSocket.OPEN) {  
-                const message = JSON.stringify({  
-                    options: { serviceId: 'OpenAITranscriptionService' },  
-                    audioData: base64Data  
-                });  
-                webSocketRef.current.send(message);  
-            }  
-        } catch (error) {  
-            setError(`Error processing audio: ${error.message}`);  
-        } finally {  
-            stream.getTracks().forEach(track => track.stop());  
-        }  
-    }, [convertBlobToBase64, setError]);
 
     const startRecording = useCallback(async () => {  
         try {  
@@ -77,24 +51,38 @@ const AudioRecorder = ({ setInput, isStreaming, setError }) => {
             webSocketRef.current = ws;
 
             const stream = await navigator.mediaDevices.getUserMedia({ audio: true });  
-            mediaRecorderRef.current = new MediaRecorder(stream, { mimeType: 'audio/webm' });  
-              
+            mediaRecorderRef.current = new MediaRecorder(stream);  
+            const audioChunks = [];
+
             mediaRecorderRef.current.ondataavailable = (event) => {  
-                audioChunksRef.current.push(event.data);  
+                audioChunks.push(event.data);  
             };
 
-            mediaRecorderRef.current.onstart = () => setIsRecording(true);  
-            mediaRecorderRef.current.onstop = () => handleRecordingStop(stream);
+            mediaRecorderRef.current.onstop = async () => {  
+                setIsRecording(false);  
+                const audioBlob = new Blob(audioChunks);  
+                const base64Data = await convertBlobToBase64(audioBlob);
+
+                if (webSocketRef.current && webSocketRef.current.readyState === WebSocket.OPEN) {  
+                    webSocketRef.current.send(JSON.stringify({  
+                        options: { serviceId: 'OpenAITranscriptionService' },  
+                        audioData: base64Data  
+                    }));  
+                }
+
+                stream.getTracks().forEach(track => track.stop());  
+            };
 
             mediaRecorderRef.current.start();  
+            setIsRecording(true);  
         } catch (error) {  
-            setError(`Error starting recording: ${error.message}`);  
-            console.error('Recording error:', error);  
+            setError(`Error starting recording or WebSocket connection: ${error.message}`);  
+            console.error("Error starting recording:", error);  
         }  
-    }, [initializeWebSocket, handleRecordingStop, setError]);
+    }, [initializeWebSocket, convertBlobToBase64, setError]);
 
     const stopRecording = useCallback(() => {  
-        if (mediaRecorderRef.current?.state !== 'inactive') {  
+        if (mediaRecorderRef.current && mediaRecorderRef.current.state !== 'inactive') {  
             mediaRecorderRef.current.stop();  
         }  
     }, []);
@@ -104,20 +92,19 @@ const AudioRecorder = ({ setInput, isStreaming, setError }) => {
             if (webSocketRef.current) {  
                 webSocketRef.current.close();  
             }  
-            if (mediaRecorderRef.current?.state !== 'inactive') {  
-                mediaRecorderRef.current.stop();  
-            }  
         };  
     }, []);
 
+    const buttonText = isRecording ? t("stop_recording") : t("start_recording");  
+    const titleText = isRecording ? t("stop_recording_title") : t("start_recording_title");
+
     return (  
-        <button   
-            className={`audio-recorder-button ${isRecording ? 'recording' : ''}`}  
-            title={isRecording ? t('stop_recording_title') : t('start_recording_title')}  
+        <button  
+            title={titleText}  
             disabled={isStreaming}  
             onClick={isRecording ? stopRecording : startRecording}  
         >  
-            {isRecording ? t('stop_recording') : t('start_recording')}  
+            {buttonText}  
         </button>  
     );  
 };
@@ -128,4 +115,4 @@ AudioRecorder.propTypes = {
     setError: PropTypes.func.isRequired  
 };
 
-export default AudioRecorder;  
+export default AudioRecorder;
