--- frontend/src/components/layout/Conversation/ConversationFooter/AudioRecorder.scss
@@ -1,32 +1,0 @@
-// AudioRecorder.scss  
-.audio-recorder-button {  
-    padding: 10px 20px;  
-    border-radius: 4px;  
-    border: none;  
-    cursor: pointer;  
-    font-size: 16px;  
-    transition: all 0.3s ease;  
-      
-    &:disabled {  
-        opacity: 0.5;  
-        cursor: not-allowed;  
-    }  
-      
-    &.recording {  
-        background-color: #ff4444;  
-        color: white;  
-          
-        &:hover:not(:disabled) {  
-            background-color: #ff2222;  
-        }  
-    }  
-      
-    &:not(.recording) {  
-        background-color: #4CAF50;  
-        color: white;  
-          
-        &:hover:not(:disabled) {  
-            background-color: #45a049;  
-        }  
-    }  
-}
